web_app/requirements.txt
flask==2.3.3
flask-cors==4.0.0
anthropic==0.8.1
python-dotenv==1.0.0
werkzeug==2.3.7
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.1
gunicorn==21.2.0 